Minor tweaks to support new XML parser; AssignmentNode priority is now an integer as spec'd in the schema
git-svn-id: https://svn.code.sf.net/p/plexil/code/branches/plexil-4@3956 d2be67b0-814c-0410-949f-d215b6b29573

+++ src/exec/Node.hh
@@ -194,7 +194,7 @@ namespace PLEXIL {
      * @return the priority of this node.
      * @note Default method; only assignment nodes care about priority.
      */
-    virtual double getPriority() const {return WORST_PRIORITY;}
+    virtual int32_t getPriority() const {return WORST_PRIORITY;}
 
     /**
      * @brief Gets the destination state of this node, were it to transition, based on the values of various conditions.
